+++ src/main/java/wooteco/subway/controller/dto/ControllerDtoAssembler.java
@@ -6,8 +6,8 @@ import wooteco.subway.controller.dto.path.PathRequest;
 import wooteco.subway.controller.dto.path.PathResponse;
 import wooteco.subway.controller.dto.section.SectionRequest;
 import wooteco.subway.controller.dto.station.StationResponse;
-import wooteco.subway.service.dto.PathRequestDto;
-import wooteco.subway.service.dto.PathResponseDto;
+import wooteco.subway.service.dto.path.PathRequestDto;
+import wooteco.subway.service.dto.path.PathResponseDto;
 import wooteco.subway.service.dto.line.LineRequestDto;
 import wooteco.subway.service.dto.line.LineResponseDto;
 import wooteco.subway.service.dto.section.SectionRequestDto;
