+++ src/main/java/wooteco/subway/controller/dto/line/LineRequest.java
@@ -2,7 +2,6 @@ package wooteco.subway.controller.dto.line;
 
 public class LineRequest {
 
-    public static final long DEFAULT_VALUE = 0L;
     private String name;
     private String color;
     private Long upStationId;
@@ -10,10 +9,6 @@ public class LineRequest {
     private int distance;
 
     public LineRequest() {
-    }
-
-    public LineRequest(String name, String color) {
-        this(name, color, DEFAULT_VALUE, DEFAULT_VALUE, 0);
     }
 
     public LineRequest(String name, String color, Long upStationId, Long downStationId, int distance) {
