+++ src/main/java/wooteco/subway/controller/dto/line/LineResponse.java
@@ -1,9 +1,7 @@
 package wooteco.subway.controller.dto.line;
 
 import wooteco.subway.controller.dto.station.StationResponse;
-import wooteco.subway.domain.Line;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class LineResponse {
@@ -13,20 +11,6 @@ public class LineResponse {
     private List<StationResponse> stations;
 
     private LineResponse() {
-    }
-
-    public LineResponse(Line line, List<StationResponse> stations) {
-        this.id = line.getId();
-        this.name = line.getName();
-        this.color = line.getColor();
-        this.stations = stations;
-    }
-
-    public LineResponse(Long id, String name, String color) {
-        this.id = id;
-        this.name = name;
-        this.color = color;
-        this.stations = new ArrayList<>();
     }
 
     public LineResponse(Long id, String name, String color, List<StationResponse> stations) {
