+++ src/main/java/wooteco/subway/controller/dto/section/SectionRequest.java
@@ -5,7 +5,7 @@ public class SectionRequest {
     private Long downStationId;
     private int distance;
 
-    public SectionRequest() {
+    private SectionRequest() {
     }
 
     public SectionRequest(Long upStationId, Long downStationId, int distance) {
