+++ src/main/java/wooteco/subway/controller/dto/station/StationRequest.java
@@ -1,9 +1,12 @@
 package wooteco.subway.controller.dto.station;
 
 public class StationRequest {
+
     private String name;
-    public StationRequest() {
+
+    private StationRequest() {
     }
+
     public StationRequest(String name) {
         this.name = name;
     }
