+++ src/main/java/wooteco/subway/controller/dto/station/StationResponse.java
@@ -1,11 +1,9 @@
 package wooteco.subway.controller.dto.station;
 
 public class StationResponse {
+
     private Long id;
     private String name;
-
-    public StationResponse() {
-    }
 
     public StationResponse(Long id, String name) {
         this.id = id;
@@ -18,13 +16,5 @@ public class StationResponse {
 
     public String getName() {
         return name;
-    }
-
-    @Override
-    public String toString() {
-        return "StationResponse{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                '}';
     }
 }
