+++ src/main/java/wooteco/subway/service/dto/line/LineRequestDto.java
@@ -2,7 +2,6 @@ package wooteco.subway.service.dto.line;
 
 public class LineRequestDto {
 
-    public static final long DEFAULT_VALUE = 0L;
     private String name;
     private String color;
     private Long upStationId;
@@ -10,10 +9,6 @@ public class LineRequestDto {
     private int distance;
 
     public LineRequestDto() {
-    }
-
-    public LineRequestDto(String name, String color) {
-        this(name, color, DEFAULT_VALUE, DEFAULT_VALUE, 0);
     }
 
     public LineRequestDto(String name, String color, Long upStationId, Long downStationId, int distance) {
