+++ src/main/java/wooteco/subway/service/dto/path/PathRequestDto.java
@@ -1,4 +1,4 @@
-package wooteco.subway.service.dto;
+package wooteco.subway.service.dto.path;
 
 public class PathRequestDto {
