+++ src/main/java/wooteco/subway/service/dto/station/StationResponseDto.java
@@ -30,14 +30,6 @@ public class StationResponseDto {
     }
 
     @Override
-    public String toString() {
-        return "StationResponseDto{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                '}';
-    }
-
-    @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
